requirements.txt
python-dotenv~=1.0.0
pydub~=0.25.1
tiktoken==0.5.1
openai==1.3.3
python-telegram-bot==20.3
requests~=2.31.0
tenacity==8.2.2
wolframalpha~=5.0.0
duckduckgo_search~=3.8.3
spotipy~=2.23.0
pytube~=15.0.0
gtts~=2.3.2
whois~=0.9.27
Pillow~=10.1.0
cairosvg~=2.7.1
imgkit~=1.2.3
matplotlib~=3.8.2
sympy~=1.3.0
